Update: Net 模块下新增了更多的 Socket 选项，并修复了部分已有的 Socket 选项；针对 DateTime 模块中关于格式化字符串功能，新增支持反转义符，以避免不必要的解析。

+++ src/TUI/Terminal.hpp
@@ -70,7 +70,7 @@ namespace Tiny {
             bool operator==(const Size& other) const { return (width == other.width && height == other.height); }
             bool operator!=(const Size& other) const { return (width != other.width || height != other.height); }
             bool isEqual(const Size& other) const { return width * height == other.width * other.height; }
-            int compare(const Size& other) const {
+            int8_t compare(const Size& other) const {
                 if (height < other.height) return 1;
                 if (height == other.height) {
                     return (width < other.width) ? 1 : ((width > other.width) ? -1 : 0);
@@ -115,7 +115,7 @@ namespace Tiny {
 
             bool operator==(const Position& other) const { return (row == other.row && column == other.column); }
             bool operator!=(const Position& other) const { return (row != other.row || column == other.column); }
-            int compare(const Position& other) const {
+            int8_t compare(const Position& other) const {
                 if (row < other.row) return 1;
                 if (row == other.row) {
                     return (column < other.column) ? 1 : ((column > other.column) ? -1 : 0);
